Fixed some bugs
The error of getting first_article via hd has been fixed, where they tried to take the first element without checking (Now we are not trying). In addition, the message that there are no articles looks more or less human (i.e. the text of the notification has been corrected)